chore: rename parse_instructions files to instructions
